view and validations
