--- src/main/common/form/models/Interest/interest.ts
@@ -1,5 +1,0 @@
-import {InterestStartDate} from '../Interest/interestStartDate';
-
-export class Interest {
-  interestStartDate?: InterestStartDate;
-}
