+++ src/main/common/form/models/interest/interest.ts
@@ -1,0 +1,9 @@
+import {InterestStartDate} from './interestStartDate';
+
+export class Interest {
+  interestStartDate?: InterestStartDate;
+
+  constructor(interestStartDate?: InterestStartDate) {
+    this.interestStartDate = interestStartDate;
+  }
+}
