+++ src/main/modules/i18n/locales/en.json
@@ -764,6 +764,11 @@
       "PAGE_TITLE": "Are you disabled?",
       "TITLE": "Are you disabled?"
     },
+    "INTEREST_START_DATE": {
+      "PAGE_TITLE": "claim interest from",
+      "TITLE": "Enter the date you want to claim interest from",
+      "REASON": "Explain why you’re claiming from this date"
+    },
     "ELIGIBILITY_APPLY_FOR_HELP_WITH_FEES": {
       "PARAGRAPH1_1": "If you have already applied for Help with Fees in respect of",
       "PARAGRAPH1_2": "you may already have a reference number. If so, you can save and continue and enter it when asked. Do not use a number related to a different claim. If you don’t have a Help with Fees number for this claim, please follow the instructions below.",
@@ -839,9 +844,9 @@
       "HELP_WITH_FEES_INTRO": "Making an application for Help with Fees does not guarantee that you will get your fee covered. You may still have to pay some or all of your court fee to get the claim issued (for the court system to officially start the claim and send details to the other side). When you apply for Help with Fees, you should receive an update from us within 5 working days, giving you the result of your application.",
       "TITLE": "Some useful information about Help with Fees"
     },
-    "EMPLOYER" : {
-      "EMPLOYER_NAME" : "Employer name",
-      "JOB_TITLE" : "Job title"
+    "EMPLOYER": {
+      "EMPLOYER_NAME": "Employer name",
+      "JOB_TITLE": "Job title"
     },
     "EMPLOYMENT_STATUS": {
       "BOTH": "You can select both.",
@@ -1772,6 +1777,8 @@
     "VALID_VALUE": "Are you sure this is a valid value?",
     "VALID_YEAR": "Enter a valid year",
     "VALID_YES_NO_OPTION": "Choose option: Yes or No",
-    "VALID_YES_NO_SELECTION": "Please select yes or no"
+    "VALID_YES_NO_SELECTION": "Please select yes or no",
+    "VALID_WHY_FROM_PARTICULAR_DATE": "Enter text explaining why do you disagree",
+    "VALID_WHY_FROM_PARTICULAR_DATE": "Enter text explaining why you're claiming from a particular date"
   }
 }
