+++ src/main/routes/routes.ts
@@ -134,7 +134,7 @@ import phoneOrVideoHearingController from './features/directionsQuestionnaire/ph
 import cantAttendHearingInNext12MonthsController
   from './features/directionsQuestionnaire/cantAttendHearingInNext12MonthsController';
 import youCanUseServiceController from './features/public/eligibility/youCanUseServiceController';
-import interestStartDateController from './features/interest/interestStartDateController';
+import interestStartDateController from './features/claim/interest/interestStartDateController';
 
 export default [
   homeController,
