+++ src/main/services/features/claim/interest/interestService.ts
@@ -1,5 +1,5 @@
 import {getCaseDataFromStore, saveDraftClaim} from '../../../../modules/draft-store/draftStoreService';
-import {Interest} from 'common/form/models/Interest/interest';
+import {Interest} from 'common/form/models/interest/interest';
 
 const {Logger} = require('@hmcts/nodejs-logging');
 const logger = Logger.getLogger('directionQuestionnaireService');
@@ -18,7 +18,7 @@ const getInterest = async (claimId: string): Promise<Interest> => {
 const saveInterest = async (claimId: string, value: any, interestPropertyName: string): Promise<void> => {
   try {
     const claim: any = await getCaseDataFromStore(claimId);
-    if (claim.interest[interestPropertyName]) {
+    if (claim.interest) {
       claim.interest[interestPropertyName] = value;
     } else {
       const interest: any = new Interest();
