+++ src/main/views/features/claim/interest/interest-start-date.njk
@@ -1,0 +1,105 @@
+{% extends "claim-details-tpl.njk" %}
+{% from "govuk/components/button/macro.njk" import govukButton %}
+{% from "govuk/components/date-input/macro.njk" import govukDateInput %}
+{% from "govuk/components/textarea/macro.njk" import govukTextarea %}
+{% from "../../../macro/contact-us-for-help.njk" import contactUsForHelp %}
+{% from "../../../macro/csrf.njk" import csrfProtection %}
+{% from "../../../macro/page-title.njk" import setPageTitle %}
+{% from "../../../macro/error-summary.njk" import errorSummary %}
+
+{% block pageTitle %}
+  {% if claimantView %}
+    {{ setPageTitle('PAGES.INTEREST_START_DATE.PAGE_TITLE') }}
+  {% else %}
+    {{ setPageTitle('PAGES.INTEREST_START_DATE.PAGE_TITLE') }}
+  {% endif %}
+{% endblock %}
+
+{% block content %}
+  <div class="govuk-width-container">
+    <main class="govuk-main-wrapper" role="main">
+      {% set errors = form.getErrors() %}
+      {% if form.hasErrors() %}
+        {{ errorSummary(errors, t) }}
+      {% endif %}
+      <div class="govuk-grid-row">
+        <div class="govuk-grid-column-two-thirds">
+          <h1 class="govuk-heading-l">
+              {{ t('PAGES.INTEREST_START_DATE.TITLE') }}
+          </h1>
+        </div>
+      </div>
+      <div class="govuk-grid-row">
+        <div class="govuk-grid-column-two-thirds">
+          <form method="post" novalidate>
+            {{ csrfProtection(csrf) }}
+            <div class="govuk-form-group {% if form.hasErrors() %} govuk-form-group--error{% endif %}">
+              {% if form.hasErrors() %}
+                {% for error in errors %}
+                  <span class="govuk-error-message">{{ t(error.text) }}</span>
+                {% endfor %}
+              {% endif %}
+              {{ govukDateInput({
+                id: "dateOfBirth",
+                type: "number",
+                inputmode: "numeric",
+                name: "dateOfBirth",
+                hint: {
+                  text: t('COMMON.DATE_EXAMPLE', {
+                    date: today.getDate() + ' ' + today.getMonth() + ' ' + (today.getFullYear() - 18)
+                  })
+                },
+                items: [
+                  {
+                    classes: "govuk-input--width-2 govuk-input--error" if form.hasFieldError('day') else "govuk-input--width-2",
+                    id: "day",
+                    label: t('COMMON.DAY'),
+                    name: "day",
+                    value: form.model.day,
+                    max: 31
+                  },
+                  {
+                    classes:"govuk-input--width-2 govuk-input--error" if form.hasFieldError('month') else "govuk-input--width-2",
+                    id: "month",
+                    label: t('COMMON.MONTH'),
+                    name: "month",
+                    value: form.model.month,
+                    max: 12
+                  },
+                  {
+                    classes: "govuk-input--width-4 govuk-input--error" if form.hasFieldError('year') else "govuk-input--width-4",
+                    id: "year",
+                    label: t('COMMON.YEAR'),
+                    name: "year",
+                    value: form.model.year,
+                    max: 9999
+                  }
+                ]
+              }) }}
+              {{ govukTextarea({
+                name: "reason",
+                id: "reason",
+                value: form.model.reason,
+                classes:"govuk-grid-column-two-thirds",
+                label: {
+                  classes: "govuk-label--s",
+                  html: t('PAGES.INTEREST_START_DATE.REASON')
+                },
+                errorMessage: {
+                  text: t(form.errorFor('reason'))
+                } if form.hasFieldError('reason')
+              }) }}
+            </div>
+            {{ govukButton({ text: t("COMMON.BUTTONS.SAVE_AND_CONTINUE") }) }}
+          </form>
+        </div>
+      </div>
+      <div class="govuk-grid-row">
+        <div class="govuk-grid-column-three-thirds">
+          {{ contactUsForHelp(t) }}
+        </div>
+      </div>
+    </main>
+  </div>
+{% endblock %}
+
